first tested version: getting values from sensors TODO: - MPU value calculations - APDS value calculation - IRQ handle thread

+++ Main_UserSpace_App/Main_UserSpace_App/InfluxDB.cpp
@@ -47,7 +47,7 @@ void InfluxDB::Set_Measurement(std::string const &Measurement)
 }
 
 
-void InfluxDB::Set_Timestamp(uint32_t Timestamp)
+void InfluxDB::Set_Timestamp(int64_t Timestamp)
 {
 	mTimestamp = Timestamp;
 }
