+++ Main_UserSpace_App/Main_UserSpace_App/InfluxDB.h
@@ -15,7 +15,7 @@ public:
 	InfluxDB(TAG_MAP const &tag_set, FIELD_MAP const &field_set, int32_t timestamp = -1, std::string const &Measurement = "");
 	
 	// METHODs
-	void				Set_Timestamp(uint32_t Timestamp);
+	void				Set_Timestamp(int64_t Timestamp);
 	void				Set_Measurement(std::string const &Measurement);
 	std::string const   Get_RawString()		const;
 private:	
@@ -23,6 +23,6 @@ private:
 	std::string		 mMeasurement;
 	const TAG_MAP	&mTag_set;
 	const FIELD_MAP	&mField_set;
-	int32_t			 mTimestamp; 	// set to -1 if default should be used
+	int64_t			 mTimestamp; 	// set to -1 if default should be used
 };
